trunk/ (mint@mint) BUG FIX: starting to fix compiler offsets on 64-bit Linux machines. Fixed compoud offset assignment for generic types, which assumed that sizeof(int) == sizeof(ssize_t). Several list tests pass again, but two still fail.
git-svn-id: svn+ssh://saulius.grazulis.lt/home/saulius/svn-repositories/compilers/trunk@2969 f7c8cbe2-f9e6-0310-9050-d7f47bf15973

+++ sc/dnode.h
@@ -111,15 +111,15 @@ ssize_t dnode_offset( DNODE *dnode );
 
 TNODE *dnode_type( DNODE *dnode );
 
-DNODE *dnode_set_offset( DNODE *dnode, int offset );
+DNODE *dnode_set_offset( DNODE *dnode, ssize_t offset );
 
-DNODE *dnode_update_offset( DNODE *dnode, int offset );
+DNODE *dnode_update_offset( DNODE *dnode, ssize_t offset );
 
-void dnode_assign_offset( DNODE *dnode, int *offset );
+void dnode_assign_offset( DNODE *dnode, ssize_t *offset );
 
-void dnode_list_assign_offsets( DNODE *dnode_list, int *offset );
+void dnode_list_assign_offsets( DNODE *dnode_list, ssize_t *offset );
 
-DNODE *dnode_set_offset( DNODE *dnode, int offset );
+DNODE *dnode_set_offset( DNODE *dnode, ssize_t offset );
 
 DNODE *dnode_insert_type( DNODE *dnode, TNODE *tnode );
